+++ tests/trace/projects/ep/potato.m
@@ -1,0 +1,3 @@
+function potato
+    disp(foo() + bar());
+end