+++ tests/trace/projects/lib_a/foo.m
@@ -1,0 +1,4 @@
+function x = foo()
+    %| pragma Tag("footag");
+    x = 0;
+end