+++ tests/trace/projects/lib_b/bar.m
@@ -1,0 +1,4 @@
+function x = bar()
+    %| pragma Tag("bartag");
+    x = 0;
+end